Added frontent UI

+++ backend/controllers/goalController.js
@@ -1,12 +1,13 @@
 const asyncHandler = require('express-async-handler')
 
 const Goal = require('../model/goalModel')
+const User = require('../model/userModel')
 
 //@desc Get goals
 //@route GET /api/goals
 //@access Private
 const getGoals = asyncHandler ( async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({ user: req.user.id })
 
     res.status(200).json(goals)
 })
@@ -21,7 +22,8 @@ const setGoal =  asyncHandler (async (req, res) => {
     }
 
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id,
     })
 
     res.status(200).json(goal)
@@ -37,7 +39,15 @@ const updateGoals = asyncHandler ( async (req, res) => {
         res.status(400)
         throw new Error('Goal not found')
     }
-
+    const user = await User.findById(req.user.id)
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+    if(goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
+    }
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, { new: true})
     res.status(200).json(updatedGoal)
 })
@@ -51,6 +61,15 @@ const deleteGoal = asyncHandler ( async (req, res) => {
     if(!goal) {
         res.status(400)
         throw new Error('Goal does not exist')
+    }
+    const user = await User.findById(req.user.id)
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+    if(goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     const deletedGoal = await Goal.findByIdAndDelete(req.params.id)
